separated the solar creation modal
